# PERF: Fixed perf regression in TimedeltaIndex.get_loc