Closes https://github.com/pandas-dev/pandas/issues/34510

1.0.3

```python
In [10]: index = pd.timedelta_range(start="1985", periods=1000, freq="D")
    ...: timedelta = index[500]
    ...:
    ...: %timeit index.get_loc(timedelta)

3.7 µs ± 275 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

```

This PR

```
3.25 µs ± 67.7 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```